requirements.txt
Flask==2.0.1
Flask_Cors==3.0.10
Pillow==9.2.0
